+++ app/src/core/workItems/index.ts
@@ -1,0 +1,2 @@
+export * from './workItemStore'
+export * from './workItemTypes'